stake-pool: Split up tests to avoid CI failures
