+++ stake-pool/program/tests/withdraw.rs
@@ -4,14 +4,13 @@
 mod helpers;
 
 use {
-    bincode::deserialize,
     borsh::BorshSerialize,
     helpers::*,
     solana_program::{
         borsh::try_from_slice_unchecked,
         instruction::{AccountMeta, Instruction, InstructionError},
         pubkey::Pubkey,
-        stake, sysvar,
+        sysvar,
     },
     solana_program_test::*,
     solana_sdk::{
@@ -19,10 +18,7 @@ use {
         transaction::{Transaction, TransactionError},
         transport::TransportError,
     },
-    spl_stake_pool::{
-        error::StakePoolError, id, instruction, minimum_stake_lamports, state,
-        MINIMUM_RESERVE_LAMPORTS,
-    },
+    spl_stake_pool::{error::StakePoolError, id, instruction, state, MINIMUM_RESERVE_LAMPORTS},
     spl_token::error::TokenError,
     test_case::test_case,
 };
@@ -829,826 +825,5 @@ async fn fail_with_not_enough_tokens() {
             0,
             InstructionError::Custom(TokenError::InsufficientFunds as u32),
         )
-    );
-}
-
-#[tokio::test]
-async fn fail_remove_validator() {
-    let (
-        mut context,
-        stake_pool_accounts,
-        validator_stake,
-        deposit_info,
-        user_transfer_authority,
-        user_stake_recipient,
-        _,
-    ) = setup(spl_token::id()).await;
-
-    // decrease a little stake, not all
-    let error = stake_pool_accounts
-        .decrease_validator_stake(
-            &mut context.banks_client,
-            &context.payer,
-            &context.last_blockhash,
-            &validator_stake.stake_account,
-            &validator_stake.transient_stake_account,
-            deposit_info.stake_lamports / 2,
-            validator_stake.transient_stake_seed,
-        )
-        .await;
-    assert!(error.is_none());
-
-    // warp forward to deactivation
-    let first_normal_slot = context.genesis_config().epoch_schedule.first_normal_slot;
-    let slots_per_epoch = context.genesis_config().epoch_schedule.slots_per_epoch;
-    context
-        .warp_to_slot(first_normal_slot + slots_per_epoch)
-        .unwrap();
-
-    // update to merge deactivated stake into reserve
-    stake_pool_accounts
-        .update_all(
-            &mut context.banks_client,
-            &context.payer,
-            &context.last_blockhash,
-            &[validator_stake.vote.pubkey()],
-            false,
-        )
-        .await;
-
-    // Withdraw entire account, fail because some stake left
-    let validator_stake_account =
-        get_account(&mut context.banks_client, &validator_stake.stake_account).await;
-    let remaining_lamports = validator_stake_account.lamports;
-    let new_user_authority = Pubkey::new_unique();
-    let error = stake_pool_accounts
-        .withdraw_stake(
-            &mut context.banks_client,
-            &context.payer,
-            &context.last_blockhash,
-            &user_stake_recipient.pubkey(),
-            &user_transfer_authority,
-            &deposit_info.pool_account.pubkey(),
-            &validator_stake.stake_account,
-            &new_user_authority,
-            remaining_lamports,
-        )
-        .await
-        .unwrap()
-        .unwrap();
-    assert_eq!(
-        error,
-        TransactionError::InstructionError(
-            0,
-            InstructionError::Custom(StakePoolError::StakeLamportsNotEqualToMinimum as u32)
-        )
-    );
-}
-
-#[tokio::test]
-async fn success_remove_validator() {
-    let (
-        mut context,
-        stake_pool_accounts,
-        validator_stake,
-        deposit_info,
-        user_transfer_authority,
-        user_stake_recipient,
-        _,
-    ) = setup(spl_token::id()).await;
-
-    let rent = context.banks_client.get_rent().await.unwrap();
-    let stake_rent = rent.minimum_balance(std::mem::size_of::<stake::state::StakeState>());
-
-    // decrease all of stake
-    let error = stake_pool_accounts
-        .decrease_validator_stake(
-            &mut context.banks_client,
-            &context.payer,
-            &context.last_blockhash,
-            &validator_stake.stake_account,
-            &validator_stake.transient_stake_account,
-            deposit_info.stake_lamports + stake_rent,
-            validator_stake.transient_stake_seed,
-        )
-        .await;
-    assert!(error.is_none());
-
-    // warp forward to deactivation
-    let first_normal_slot = context.genesis_config().epoch_schedule.first_normal_slot;
-    let slots_per_epoch = context.genesis_config().epoch_schedule.slots_per_epoch;
-    context
-        .warp_to_slot(first_normal_slot + slots_per_epoch)
-        .unwrap();
-
-    // update to merge deactivated stake into reserve
-    stake_pool_accounts
-        .update_all(
-            &mut context.banks_client,
-            &context.payer,
-            &context.last_blockhash,
-            &[validator_stake.vote.pubkey()],
-            false,
-        )
-        .await;
-
-    let validator_stake_account =
-        get_account(&mut context.banks_client, &validator_stake.stake_account).await;
-    let remaining_lamports = validator_stake_account.lamports;
-    let new_user_authority = Pubkey::new_unique();
-    let pool_tokens = stake_pool_accounts.calculate_inverse_withdrawal_fee(remaining_lamports);
-    let error = stake_pool_accounts
-        .withdraw_stake(
-            &mut context.banks_client,
-            &context.payer,
-            &context.last_blockhash,
-            &user_stake_recipient.pubkey(),
-            &user_transfer_authority,
-            &deposit_info.pool_account.pubkey(),
-            &validator_stake.stake_account,
-            &new_user_authority,
-            pool_tokens,
-        )
-        .await;
-    assert!(error.is_none());
-
-    // Check validator stake account gone
-    let validator_stake_account = context
-        .banks_client
-        .get_account(validator_stake.stake_account)
-        .await
-        .unwrap();
-    assert!(validator_stake_account.is_none());
-
-    // Check user recipient stake account balance
-    let user_stake_recipient_account =
-        get_account(&mut context.banks_client, &user_stake_recipient.pubkey()).await;
-    assert_eq!(
-        user_stake_recipient_account.lamports,
-        remaining_lamports + stake_rent + 1
-    );
-
-    // Check that cleanup happens correctly
-    stake_pool_accounts
-        .cleanup_removed_validator_entries(
-            &mut context.banks_client,
-            &context.payer,
-            &context.last_blockhash,
-        )
-        .await;
-
-    let validator_list = get_account(
-        &mut context.banks_client,
-        &stake_pool_accounts.validator_list.pubkey(),
-    )
-    .await;
-    let validator_list =
-        try_from_slice_unchecked::<state::ValidatorList>(validator_list.data.as_slice()).unwrap();
-    let validator_stake_item = validator_list.find(&validator_stake.vote.pubkey());
-    assert!(validator_stake_item.is_none());
-}
-
-#[tokio::test]
-async fn fail_with_reserve() {
-    let (
-        mut context,
-        stake_pool_accounts,
-        validator_stake,
-        deposit_info,
-        user_transfer_authority,
-        user_stake_recipient,
-        tokens_to_burn,
-    ) = setup(spl_token::id()).await;
-
-    // decrease a little stake, not all
-    let error = stake_pool_accounts
-        .decrease_validator_stake(
-            &mut context.banks_client,
-            &context.payer,
-            &context.last_blockhash,
-            &validator_stake.stake_account,
-            &validator_stake.transient_stake_account,
-            deposit_info.stake_lamports / 2,
-            validator_stake.transient_stake_seed,
-        )
-        .await;
-    assert!(error.is_none());
-
-    // warp forward to deactivation
-    let first_normal_slot = context.genesis_config().epoch_schedule.first_normal_slot;
-    let slots_per_epoch = context.genesis_config().epoch_schedule.slots_per_epoch;
-    context
-        .warp_to_slot(first_normal_slot + slots_per_epoch)
-        .unwrap();
-
-    // update to merge deactivated stake into reserve
-    stake_pool_accounts
-        .update_all(
-            &mut context.banks_client,
-            &context.payer,
-            &context.last_blockhash,
-            &[validator_stake.vote.pubkey()],
-            false,
-        )
-        .await;
-
-    // Withdraw directly from reserve, fail because some stake left
-    let new_user_authority = Pubkey::new_unique();
-    let error = stake_pool_accounts
-        .withdraw_stake(
-            &mut context.banks_client,
-            &context.payer,
-            &context.last_blockhash,
-            &user_stake_recipient.pubkey(),
-            &user_transfer_authority,
-            &deposit_info.pool_account.pubkey(),
-            &stake_pool_accounts.reserve_stake.pubkey(),
-            &new_user_authority,
-            tokens_to_burn,
-        )
-        .await
-        .unwrap()
-        .unwrap();
-    assert_eq!(
-        error,
-        TransactionError::InstructionError(
-            0,
-            InstructionError::Custom(StakePoolError::StakeLamportsNotEqualToMinimum as u32)
-        )
-    );
-}
-
-#[tokio::test]
-async fn success_with_reserve() {
-    let (
-        mut context,
-        stake_pool_accounts,
-        validator_stake,
-        deposit_info,
-        user_transfer_authority,
-        user_stake_recipient,
-        _,
-    ) = setup(spl_token::id()).await;
-
-    let rent = context.banks_client.get_rent().await.unwrap();
-    let stake_rent = rent.minimum_balance(std::mem::size_of::<stake::state::StakeState>());
-
-    // decrease all of stake
-    let error = stake_pool_accounts
-        .decrease_validator_stake(
-            &mut context.banks_client,
-            &context.payer,
-            &context.last_blockhash,
-            &validator_stake.stake_account,
-            &validator_stake.transient_stake_account,
-            deposit_info.stake_lamports + stake_rent,
-            validator_stake.transient_stake_seed,
-        )
-        .await;
-    assert!(error.is_none());
-
-    // warp forward to deactivation
-    let first_normal_slot = context.genesis_config().epoch_schedule.first_normal_slot;
-    let slots_per_epoch = context.genesis_config().epoch_schedule.slots_per_epoch;
-    context
-        .warp_to_slot(first_normal_slot + slots_per_epoch)
-        .unwrap();
-
-    // update to merge deactivated stake into reserve
-    stake_pool_accounts
-        .update_all(
-            &mut context.banks_client,
-            &context.payer,
-            &context.last_blockhash,
-            &[validator_stake.vote.pubkey()],
-            false,
-        )
-        .await;
-
-    // now it works
-    let new_user_authority = Pubkey::new_unique();
-    let error = stake_pool_accounts
-        .withdraw_stake(
-            &mut context.banks_client,
-            &context.payer,
-            &context.last_blockhash,
-            &user_stake_recipient.pubkey(),
-            &user_transfer_authority,
-            &deposit_info.pool_account.pubkey(),
-            &stake_pool_accounts.reserve_stake.pubkey(),
-            &new_user_authority,
-            deposit_info.pool_tokens,
-        )
-        .await;
-    assert!(error.is_none());
-
-    // first and only deposit, lamports:pool 1:1
-    let stake_pool = get_account(
-        &mut context.banks_client,
-        &stake_pool_accounts.stake_pool.pubkey(),
-    )
-    .await;
-    let stake_pool =
-        try_from_slice_unchecked::<state::StakePool>(stake_pool.data.as_slice()).unwrap();
-    // the entire deposit is actually stake since it isn't activated, so only
-    // the stake deposit fee is charged
-    let deposit_fee = stake_pool
-        .calc_pool_tokens_stake_deposit_fee(stake_rent + deposit_info.stake_lamports)
-        .unwrap();
-    assert_eq!(
-        deposit_info.stake_lamports + stake_rent - deposit_fee,
-        deposit_info.pool_tokens,
-        "stake {} rent {} deposit fee {} pool tokens {}",
-        deposit_info.stake_lamports,
-        stake_rent,
-        deposit_fee,
-        deposit_info.pool_tokens
-    );
-
-    let withdrawal_fee = stake_pool_accounts.calculate_withdrawal_fee(deposit_info.pool_tokens);
-
-    // Check tokens used
-    let user_token_balance = get_token_balance(
-        &mut context.banks_client,
-        &deposit_info.pool_account.pubkey(),
-    )
-    .await;
-    assert_eq!(user_token_balance, 0);
-
-    // Check reserve stake account balance
-    let reserve_stake_account = get_account(
-        &mut context.banks_client,
-        &stake_pool_accounts.reserve_stake.pubkey(),
-    )
-    .await;
-    let stake_state = deserialize::<stake::state::StakeState>(&reserve_stake_account.data).unwrap();
-    let meta = stake_state.meta().unwrap();
-    assert_eq!(
-        MINIMUM_RESERVE_LAMPORTS + meta.rent_exempt_reserve + withdrawal_fee + deposit_fee,
-        reserve_stake_account.lamports
-    );
-
-    // Check user recipient stake account balance
-    let user_stake_recipient_account =
-        get_account(&mut context.banks_client, &user_stake_recipient.pubkey()).await;
-    assert_eq!(
-        user_stake_recipient_account.lamports,
-        MINIMUM_RESERVE_LAMPORTS + deposit_info.stake_lamports + stake_rent * 2
-            - withdrawal_fee
-            - deposit_fee
-    );
-}
-
-#[tokio::test]
-async fn success_and_fail_with_preferred_withdraw() {
-    let (
-        mut context,
-        stake_pool_accounts,
-        validator_stake,
-        deposit_info,
-        user_transfer_authority,
-        user_stake_recipient,
-        tokens_to_burn,
-    ) = setup(spl_token::id()).await;
-
-    let preferred_validator = simple_add_validator_to_pool(
-        &mut context.banks_client,
-        &context.payer,
-        &context.last_blockhash,
-        &stake_pool_accounts,
-        None,
-    )
-    .await;
-
-    stake_pool_accounts
-        .set_preferred_validator(
-            &mut context.banks_client,
-            &context.payer,
-            &context.last_blockhash,
-            instruction::PreferredValidatorType::Withdraw,
-            Some(preferred_validator.vote.pubkey()),
-        )
-        .await;
-
-    // preferred is empty, withdrawing from non-preferred works
-    let new_authority = Pubkey::new_unique();
-    let error = stake_pool_accounts
-        .withdraw_stake(
-            &mut context.banks_client,
-            &context.payer,
-            &context.last_blockhash,
-            &user_stake_recipient.pubkey(),
-            &user_transfer_authority,
-            &deposit_info.pool_account.pubkey(),
-            &validator_stake.stake_account,
-            &new_authority,
-            tokens_to_burn / 2,
-        )
-        .await;
-    assert!(error.is_none());
-
-    // Deposit into preferred, then fail
-    let user_stake_recipient = Keypair::new();
-    create_blank_stake_account(
-        &mut context.banks_client,
-        &context.payer,
-        &context.last_blockhash,
-        &user_stake_recipient,
-    )
-    .await;
-
-    let _preferred_deposit = simple_deposit_stake(
-        &mut context.banks_client,
-        &context.payer,
-        &context.last_blockhash,
-        &stake_pool_accounts,
-        &preferred_validator,
-        TEST_STAKE_AMOUNT,
-    )
-    .await
-    .unwrap();
-
-    let error = stake_pool_accounts
-        .withdraw_stake(
-            &mut context.banks_client,
-            &context.payer,
-            &context.last_blockhash,
-            &user_stake_recipient.pubkey(),
-            &user_transfer_authority,
-            &deposit_info.pool_account.pubkey(),
-            &validator_stake.stake_account,
-            &new_authority,
-            tokens_to_burn / 2,
-        )
-        .await
-        .unwrap()
-        .unwrap();
-    assert_eq!(
-        error,
-        TransactionError::InstructionError(
-            0,
-            InstructionError::Custom(StakePoolError::IncorrectWithdrawVoteAddress as u32)
-        )
-    );
-
-    // success from preferred
-    let error = stake_pool_accounts
-        .withdraw_stake(
-            &mut context.banks_client,
-            &context.payer,
-            &context.last_blockhash,
-            &user_stake_recipient.pubkey(),
-            &user_transfer_authority,
-            &deposit_info.pool_account.pubkey(),
-            &preferred_validator.stake_account,
-            &new_authority,
-            tokens_to_burn / 2,
-        )
-        .await;
-    assert!(error.is_none());
-}
-
-#[tokio::test]
-async fn fail_withdraw_from_transient() {
-    let (
-        mut context,
-        stake_pool_accounts,
-        validator_stake_account,
-        deposit_info,
-        user_transfer_authority,
-        user_stake_recipient,
-        tokens_to_withdraw,
-    ) = setup(spl_token::id()).await;
-
-    // add a preferred withdraw validator, keep it empty, to be sure that this works
-    let preferred_validator = simple_add_validator_to_pool(
-        &mut context.banks_client,
-        &context.payer,
-        &context.last_blockhash,
-        &stake_pool_accounts,
-        None,
-    )
-    .await;
-
-    stake_pool_accounts
-        .set_preferred_validator(
-            &mut context.banks_client,
-            &context.payer,
-            &context.last_blockhash,
-            instruction::PreferredValidatorType::Withdraw,
-            Some(preferred_validator.vote.pubkey()),
-        )
-        .await;
-
-    let rent = context.banks_client.get_rent().await.unwrap();
-    let stake_rent = rent.minimum_balance(std::mem::size_of::<stake::state::StakeState>());
-
-    // decrease to minimum stake + 1 lamport
-    let error = stake_pool_accounts
-        .decrease_validator_stake(
-            &mut context.banks_client,
-            &context.payer,
-            &context.last_blockhash,
-            &validator_stake_account.stake_account,
-            &validator_stake_account.transient_stake_account,
-            deposit_info.stake_lamports + stake_rent - 1,
-            validator_stake_account.transient_stake_seed,
-        )
-        .await;
-    assert!(error.is_none());
-
-    // fail withdrawing from transient, still a lamport in the validator stake account
-    let new_user_authority = Pubkey::new_unique();
-    let error = stake_pool_accounts
-        .withdraw_stake(
-            &mut context.banks_client,
-            &context.payer,
-            &context.last_blockhash,
-            &user_stake_recipient.pubkey(),
-            &user_transfer_authority,
-            &deposit_info.pool_account.pubkey(),
-            &validator_stake_account.transient_stake_account,
-            &new_user_authority,
-            tokens_to_withdraw,
-        )
-        .await
-        .unwrap()
-        .unwrap();
-    assert_eq!(
-        error,
-        TransactionError::InstructionError(
-            0,
-            InstructionError::Custom(StakePoolError::InvalidStakeAccountAddress as u32)
-        )
-    );
-}
-
-#[tokio::test]
-async fn success_withdraw_from_transient() {
-    let (
-        mut context,
-        stake_pool_accounts,
-        validator_stake_account,
-        deposit_info,
-        user_transfer_authority,
-        user_stake_recipient,
-        tokens_to_withdraw,
-    ) = setup(spl_token::id()).await;
-
-    // add a preferred withdraw validator, keep it empty, to be sure that this works
-    let preferred_validator = simple_add_validator_to_pool(
-        &mut context.banks_client,
-        &context.payer,
-        &context.last_blockhash,
-        &stake_pool_accounts,
-        None,
-    )
-    .await;
-
-    stake_pool_accounts
-        .set_preferred_validator(
-            &mut context.banks_client,
-            &context.payer,
-            &context.last_blockhash,
-            instruction::PreferredValidatorType::Withdraw,
-            Some(preferred_validator.vote.pubkey()),
-        )
-        .await;
-
-    let rent = context.banks_client.get_rent().await.unwrap();
-    let stake_rent = rent.minimum_balance(std::mem::size_of::<stake::state::StakeState>());
-
-    // decrease all of stake
-    let error = stake_pool_accounts
-        .decrease_validator_stake(
-            &mut context.banks_client,
-            &context.payer,
-            &context.last_blockhash,
-            &validator_stake_account.stake_account,
-            &validator_stake_account.transient_stake_account,
-            deposit_info.stake_lamports + stake_rent,
-            validator_stake_account.transient_stake_seed,
-        )
-        .await;
-    assert!(error.is_none());
-
-    // nothing left in the validator stake account (or any others), so withdrawing
-    // from the transient account is ok!
-    let new_user_authority = Pubkey::new_unique();
-    let error = stake_pool_accounts
-        .withdraw_stake(
-            &mut context.banks_client,
-            &context.payer,
-            &context.last_blockhash,
-            &user_stake_recipient.pubkey(),
-            &user_transfer_authority,
-            &deposit_info.pool_account.pubkey(),
-            &validator_stake_account.transient_stake_account,
-            &new_user_authority,
-            tokens_to_withdraw / 2,
-        )
-        .await;
-    assert!(error.is_none());
-}
-
-#[tokio::test]
-async fn success_withdraw_all_fee_tokens() {
-    let (
-        mut context,
-        stake_pool_accounts,
-        validator_stake_account,
-        deposit_info,
-        user_transfer_authority,
-        user_stake_recipient,
-        tokens_to_withdraw,
-    ) = setup(spl_token::id()).await;
-
-    // move tokens to fee account
-    transfer_spl_tokens(
-        &mut context.banks_client,
-        &context.payer,
-        &context.last_blockhash,
-        &stake_pool_accounts.token_program_id,
-        &deposit_info.pool_account.pubkey(),
-        &stake_pool_accounts.pool_mint.pubkey(),
-        &stake_pool_accounts.pool_fee_account.pubkey(),
-        &user_transfer_authority,
-        tokens_to_withdraw / 2,
-        stake_pool_accounts.pool_decimals,
-    )
-    .await;
-
-    let fee_tokens = get_token_balance(
-        &mut context.banks_client,
-        &stake_pool_accounts.pool_fee_account.pubkey(),
-    )
-    .await;
-
-    let user_transfer_authority = Keypair::new();
-    delegate_tokens(
-        &mut context.banks_client,
-        &context.payer,
-        &context.last_blockhash,
-        &stake_pool_accounts.token_program_id,
-        &stake_pool_accounts.pool_fee_account.pubkey(),
-        &stake_pool_accounts.manager,
-        &user_transfer_authority.pubkey(),
-        fee_tokens,
-    )
-    .await;
-
-    let new_authority = Pubkey::new_unique();
-    let error = stake_pool_accounts
-        .withdraw_stake(
-            &mut context.banks_client,
-            &context.payer,
-            &context.last_blockhash,
-            &user_stake_recipient.pubkey(),
-            &user_transfer_authority,
-            &stake_pool_accounts.pool_fee_account.pubkey(),
-            &validator_stake_account.stake_account,
-            &new_authority,
-            fee_tokens,
-        )
-        .await;
-    assert!(error.is_none());
-
-    // Check balance is 0
-    let fee_tokens = get_token_balance(
-        &mut context.banks_client,
-        &stake_pool_accounts.pool_fee_account.pubkey(),
-    )
-    .await;
-    assert_eq!(fee_tokens, 0);
-}
-
-#[tokio::test]
-async fn success_empty_out_stake_with_fee() {
-    let (
-        mut context,
-        stake_pool_accounts,
-        _,
-        deposit_info,
-        user_transfer_authority,
-        user_stake_recipient,
-        tokens_to_withdraw,
-    ) = setup(spl_token::id()).await;
-
-    // add another validator and deposit into it
-    let other_validator_stake_account = simple_add_validator_to_pool(
-        &mut context.banks_client,
-        &context.payer,
-        &context.last_blockhash,
-        &stake_pool_accounts,
-        None,
-    )
-    .await;
-
-    let other_deposit_info = simple_deposit_stake(
-        &mut context.banks_client,
-        &context.payer,
-        &context.last_blockhash,
-        &stake_pool_accounts,
-        &other_validator_stake_account,
-        TEST_STAKE_AMOUNT,
-    )
-    .await
-    .unwrap();
-
-    // move tokens to new account
-    transfer_spl_tokens(
-        &mut context.banks_client,
-        &context.payer,
-        &context.last_blockhash,
-        &stake_pool_accounts.token_program_id,
-        &deposit_info.pool_account.pubkey(),
-        &stake_pool_accounts.pool_mint.pubkey(),
-        &other_deposit_info.pool_account.pubkey(),
-        &user_transfer_authority,
-        tokens_to_withdraw,
-        stake_pool_accounts.pool_decimals,
-    )
-    .await;
-
-    let user_tokens = get_token_balance(
-        &mut context.banks_client,
-        &other_deposit_info.pool_account.pubkey(),
-    )
-    .await;
-
-    let user_transfer_authority = Keypair::new();
-    delegate_tokens(
-        &mut context.banks_client,
-        &context.payer,
-        &context.last_blockhash,
-        &stake_pool_accounts.token_program_id,
-        &other_deposit_info.pool_account.pubkey(),
-        &other_deposit_info.authority,
-        &user_transfer_authority.pubkey(),
-        user_tokens,
-    )
-    .await;
-
-    // calculate exactly how much to withdraw, given the fee, to get the account
-    // down to 0, using an inverse fee calculation
-    let validator_stake_account = get_account(
-        &mut context.banks_client,
-        &other_validator_stake_account.stake_account,
-    )
-    .await;
-    let stake_state =
-        deserialize::<stake::state::StakeState>(&validator_stake_account.data).unwrap();
-    let meta = stake_state.meta().unwrap();
-    let stake_minimum_delegation = stake_get_minimum_delegation(
-        &mut context.banks_client,
-        &context.payer,
-        &context.last_blockhash,
-    )
-    .await;
-    let lamports_to_withdraw =
-        validator_stake_account.lamports - minimum_stake_lamports(&meta, stake_minimum_delegation);
-    let stake_pool_account = get_account(
-        &mut context.banks_client,
-        &stake_pool_accounts.stake_pool.pubkey(),
-    )
-    .await;
-    let stake_pool =
-        try_from_slice_unchecked::<state::StakePool>(stake_pool_account.data.as_slice()).unwrap();
-    let fee = stake_pool.stake_withdrawal_fee;
-    let inverse_fee = state::Fee {
-        numerator: fee.denominator - fee.numerator,
-        denominator: fee.denominator,
-    };
-    let pool_tokens_to_withdraw =
-        lamports_to_withdraw * inverse_fee.denominator / inverse_fee.numerator;
-
-    let new_authority = Pubkey::new_unique();
-    let error = stake_pool_accounts
-        .withdraw_stake(
-            &mut context.banks_client,
-            &context.payer,
-            &context.last_blockhash,
-            &user_stake_recipient.pubkey(),
-            &user_transfer_authority,
-            &other_deposit_info.pool_account.pubkey(),
-            &other_validator_stake_account.stake_account,
-            &new_authority,
-            pool_tokens_to_withdraw,
-        )
-        .await;
-    assert!(error.is_none());
-
-    // Check balance of validator stake account is MINIMUM + rent-exemption
-    let validator_stake_account = get_account(
-        &mut context.banks_client,
-        &other_validator_stake_account.stake_account,
-    )
-    .await;
-    let stake_state =
-        deserialize::<stake::state::StakeState>(&validator_stake_account.data).unwrap();
-    let meta = stake_state.meta().unwrap();
-    assert_eq!(
-        validator_stake_account.lamports,
-        minimum_stake_lamports(&meta, stake_minimum_delegation)
     );
 }
